requirements.txt
selenium>=4.15.0
flask>=3.0.0
sqlalchemy>=2.0.0
dnspython>=2.4.0
python-whois>=0.8.0
validators>=0.22.0
beautifulsoup4>=4.12.0
requests>=2.31.0
jinja2>=3.1.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
lxml>=4.9.0
undetected-chromedriver>=3.5.0
